front/requirements.txt
google-auth
pandas
requests
streamlit
